requirements.txt
Django==1.10.5
django-cors-headers==2.0.1
django-countries==4.0
django-debug-toolbar==1.6
django-extensions==1.7.6
django-jsonfield==1.0.1
django-reversion==2.0.8
djangorestframework==3.5.3
django-modeladmin-reorder
django-rest-auth
djangorestframework-jwt
django-allauth